src/data_processing/requirements.txt
# File: requirements.txt

# PDF processing
PyPDF2==3.0.1

# HTML processing
beautifulsoup4==4.12.2

# Markdown processing
Markdown==3.4.3

# Natural Language Processing
sentence-transformers==2.2.2

# Vector Database
pinecone-client==2.2.1

# Utility libraries
numpy==1.24.3
pandas==2.0.1

# Web framework (if needed for API integration)
fastapi==0.95.2
uvicorn==0.22.0

# Environment variable management
python-dotenv==1.0.0

# Optional: for progress bars in data processing
tqdm==4.65.0

# Optional: for advanced text processing
nltk==3.8.1
spacy==3.5.2

